+++ packages/riverpod/test/src/providers/async_notifier_test.dart
@@ -1015,6 +1015,33 @@ void main() {
     });
 
     group('AsyncNotifier.future', () {
+      test('does not notify when updateShouldNotify returns false', () async {
+        final provider = factory.deferredProvider<int>(
+          (ref, _) => 0,
+          updateShouldNotify: (a, b) => false,
+        );
+        final container = ProviderContainer.test();
+        final listener = Listener<Future<int>>();
+        final notifier = container.read(provider.notifier);
+
+        container.listen(provider.future, listener.call);
+
+        notifier.state = AsyncData(42);
+        expect(
+          container.read(provider.future),
+          completion(42),
+          reason:
+              'Even though no notification happened, the state is still updated',
+        );
+
+        notifier.state = AsyncError(42, StackTrace.empty);
+        expect(container.read(provider.future), throwsA(42));
+
+        notifier.state = AsyncLoading();
+
+        verifyZeroInteractions(listener);
+      });
+
       test('can be used inside Notifier.build', () async {
         final provider = factory.simpleTestProvider<int>((ref, self) {
           return self.future;